Tyler2

+++ main_app/views.py
@@ -4,6 +4,7 @@ from .forms import *
 from django.contrib.auth import login
 from django.contrib import messages
 from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 from .models import *
 
 # Static routes
@@ -26,10 +27,12 @@ def signup(request):
         return render(request, 'registration/signup.html', context)
 
 # Profile routes
-def profile_detail(request, user_id):
-    profile = Profile.objects.get(user=user_id)
+def profile_detail(request, username):
+    user = User.objects.get(username=username)
+    profile = Profile.objects.get(user=user)
     return render(request, 'profiles/detail.html', {'profile': profile})
 
+@login_required
 def my_profile(request):
     my_profile = User.objects.get(username=request.user.username)
     posts = Post.objects.filter(user=request.user)
@@ -40,6 +43,7 @@ def my_profile(request):
     }
     return render(request,'profiles/my_profile.html', context)
 
+@login_required
 def edit_profile(request):
     profile = Profile.objects.get(user_id=request.user.id)
     if request.method == 'POST':
@@ -58,26 +62,33 @@ def post_detail(request, post_id):
     print(post.__dict__, '-----------------------HERE')
     return render(request, 'posts/detail.html', {'post': post})
 
-
+@login_required
 def edit_post(request, post_id):
     post = Post.objects.get(id=post_id)
+    if request.user == post.user:
+        if request.method == 'POST':
+            post_form = PostForm(request.POST, instance=post)
+            if post_form.is_valid():
+                updated_post = post_form.save()
+                return redirect('post_detail', updated_post.id)
 
-    if request.method == 'POST':
-        post_form = PostForm(request.POST, instance=post)
-        if post_form.is_valid():
-            updated_post = post_form.save()
-            return redirect('post_detail', updated_post.id)
-
+        else:
+            form = PostForm(instance=post)
+            context = {'form': form, 'post':post}
+            return render(request, 'posts/edit_post.html', context)
     else:
-        form = PostForm(instance=post)
-        context = {'form': form, 'post':post}
-        return render(request, 'posts/edit_post.html', context)
+        return redirect('city_index')
 
-
+@login_required
 def delete_post(request, post_id):
-    Post.objects.get(id=post_id).delete()
-
-    return redirect('my_profile')
+    post = Post.objects.get(id=post_id)
+    if request.user == post.user:
+        Post.objects.get(id=post_id).delete()
+        
+        return redirect('my_profile')
+    else:
+        # flash message error
+        return redirect('my_profile')
 
 
 # City Routes
@@ -85,6 +96,7 @@ def city_index(request):
     cities = City.objects.all()
     return render(request, 'cities/index.html', {'cities': cities})
 
+@login_required
 def city_detail(request, city_id):
     user = User.objects.get(id = request.user.id)
     city = City.objects.get(id = city_id)
@@ -93,12 +105,12 @@ def city_detail(request, city_id):
     if request.method == 'POST':
         post_form = PostForm(request.POST)
         if post_form.is_valid():
-            new_post = post_form.save()
+            new_post = post_form.save(commit=False)
             new_post.user = request.user
             new_post.city = city
             new_post.save()
-            messages.success(request, 'Successful post!')
-            return render(request, 'cities/detail.html', {'city': city, 'posts': posts})
+            messages.success(request, f'Successfully created post titled: {new_post.title}!')
+            return redirect('city_detail', city_id)
     else:
         form = PostForm()
         return render(request, 'cities/detail.html', {'city': city, 'posts': posts, 'form': form})
